Add preview job and chapter preview

+++ lecture-recording-manager/Startup.cs
@@ -48,7 +48,7 @@ namespace LectureRecordingManager
             services.Configure<FormOptions>(x =>
             {
                 x.ValueLengthLimit = int.MaxValue;
-                x.MultipartBodyLengthLimit = int.MaxValue;
+                x.MultipartBodyLengthLimit = long.MaxValue;
             });
 
             services.AddSignalR();
